+++ v3/crates/metadata-resolve/src/stages/models_graphql/order_by.rs
@@ -1,10 +1,11 @@
 use crate::Warning;
 use crate::{mk_name, QualifiedTypeName};
 use lang_graphql::ast::common as ast;
+use open_dds::commands::CommandName;
 use open_dds::types::CustomTypeName;
 
 use crate::stages::{
-    graphql_config, models, object_relationships, order_by_expressions, scalar_types,
+    commands, graphql_config, models, object_relationships, order_by_expressions, scalar_types,
 };
 use crate::types::subgraph::Qualified;
 use indexmap::IndexMap;
@@ -26,6 +27,7 @@ pub fn resolve_order_by_expression(
         object_relationships::ObjectTypeWithRelationships,
     >,
     models: &IndexMap<Qualified<ModelName>, models::Model>,
+    commands: &IndexMap<Qualified<CommandName>, commands::Command>,
     scalar_types: &BTreeMap<Qualified<CustomTypeName>, scalar_types::ScalarTypeRepresentation>,
     order_by_expressions: &mut OrderByExpressions,
     graphql_types: &mut graphql_config::GraphqlTypeNames,
@@ -65,8 +67,11 @@ pub fn resolve_order_by_expression(
                     &model.name,
                     order_by_expression_name,
                     model_source,
-                    order_by_expressions,
                     &model.data_type,
+                    order_by_expressions,
+                    object_types,
+                    models,
+                    commands,
                     issues,
                 )
             })
@@ -198,19 +203,16 @@ pub fn make_order_by_expression(
 
             // Build relationship field in filter expression only when the target_model is backed by a source, we have a
             // check for the source model during the runtime
-            if let (Some(target_source), Some(model_source), Some(target_capabilities)) = (
-                &target_model.source,
-                &model_source,
-                &relationship.target_capabilities,
-            ) {
-                if let object_relationships::RelationshipExecutionCategory::Local =
-                    object_relationships::relationship_execution_category(
-                        crate::FieldNestedness::NotNested, // we don't support nested fields in
-                        // legacy OrderByExpressions
-                        &model_source.data_connector,
-                        &target_source.data_connector,
-                        target_capabilities,
-                    )
+            if let (Some(target_source), Some(model_source)) = (&target_model.source, &model_source)
+            {
+                if order_by_expressions::validate_orderable_relationship(
+                    &model.data_type,
+                    relationship_name,
+                    order_by_expressions::OrderableFieldNestedness::NotNested, // we don't support nested fields in legacy OrderByExpressions
+                    &model_source.data_connector,
+                    &target_source.data_connector.name,
+                )
+                .is_ok()
                 {
                     // TODO(naveen): Support Array relationships in order_by when the support for aggregates is implemented
                     if open_dds::relationships::RelationshipType::Object == *relationship_type {
@@ -259,8 +261,14 @@ fn resolve_order_by_expression_for_model(
     qualified_model_name: &Qualified<ModelName>,
     order_by_expression_name: &open_dds::order_by_expression::OrderByExpressionName,
     model_source: Option<&models::ModelSource>,
-    order_by_expressions: &OrderByExpressions,
     qualified_model_object_type_name: &Qualified<CustomTypeName>,
+    order_by_expressions: &OrderByExpressions,
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
+    models: &IndexMap<Qualified<ModelName>, models::Model>,
+    commands: &IndexMap<Qualified<CommandName>, commands::Command>,
     issues: &mut Vec<Warning>,
 ) -> Result<Qualified<OrderByExpressionIdentifier>, models::ModelsError> {
     let order_by_expression_identifier = Qualified::new(
@@ -295,9 +303,13 @@ fn resolve_order_by_expression_for_model(
     if let Some(model_source) = model_source {
         validate_data_connector_compatibility(
             order_by_expression,
+            order_by_expressions::OrderableFieldNestedness::NotNested,
             model_source,
             qualified_model_name,
             order_by_expressions,
+            object_types,
+            models,
+            commands,
             issues,
         )?;
     }
@@ -307,9 +319,54 @@ fn resolve_order_by_expression_for_model(
 
 fn validate_data_connector_compatibility(
     order_by_expression: &order_by_expressions::ObjectOrderByExpression,
+    orderable_field_nestedness: order_by_expressions::OrderableFieldNestedness,
     model_source: &models::ModelSource,
     model_name: &Qualified<ModelName>,
     order_by_expressions: &OrderByExpressions,
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
+    models: &IndexMap<Qualified<ModelName>, models::Model>,
+    commands: &IndexMap<Qualified<CommandName>, commands::Command>,
+    issues: &mut Vec<Warning>,
+) -> Result<(), models::ModelsError> {
+    validate_orderable_fields(
+        order_by_expression,
+        model_source,
+        order_by_expressions,
+        model_name,
+        object_types,
+        models,
+        commands,
+        issues,
+    )?;
+
+    validate_orderable_relationships(
+        object_types,
+        order_by_expression,
+        models,
+        issues,
+        model_name,
+        commands,
+        orderable_field_nestedness,
+        model_source,
+    )?;
+
+    Ok(())
+}
+
+fn validate_orderable_fields(
+    order_by_expression: &order_by_expressions::ObjectOrderByExpression,
+    model_source: &models::ModelSource,
+    order_by_expressions: &OrderByExpressions,
+    model_name: &Qualified<ModelName>,
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
+    models: &IndexMap<Qualified<ModelName>, models::Model>,
+    commands: &IndexMap<Qualified<CommandName>, commands::Command>,
     issues: &mut Vec<Warning>,
 ) -> Result<(), models::ModelsError> {
     for (field_name, orderable_field) in &order_by_expression.orderable_fields {
@@ -331,11 +388,15 @@ fn validate_data_connector_compatibility(
                                 .order_by_expression_identifier
                                 .clone(),
                         })?;
-                    validate_nested_relationship_compatibility(
+                    validate_data_connector_compatibility(
                         nested_order_by_expression,
+                        order_by_expressions::OrderableFieldNestedness::ObjectNested,
                         model_source,
                         model_name,
                         order_by_expressions,
+                        object_types,
+                        models,
+                        commands,
                         issues,
                     )?;
                 }
@@ -359,62 +420,113 @@ fn validate_data_connector_compatibility(
     Ok(())
 }
 
-fn validate_nested_relationship_compatibility(
+fn validate_orderable_relationships(
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
     order_by_expression: &order_by_expressions::ObjectOrderByExpression,
-    model_source: &models::ModelSource,
-    model_name: &Qualified<ModelName>,
-    order_by_expressions: &OrderByExpressions,
+    models: &IndexMap<Qualified<ModelName>, models::Model>,
     issues: &mut Vec<Warning>,
+    model_name: &Qualified<ModelName>,
+    commands: &IndexMap<Qualified<CommandName>, commands::Command>,
+    orderable_field_nestedness: order_by_expressions::OrderableFieldNestedness,
+    model_source: &models::ModelSource,
 ) -> Result<(), models::ModelsError> {
-    let connector_supports_ordering_nested_relationships = model_source
-        .data_connector
-        .capabilities
-        .supports_relationships
-        .as_ref()
-        .is_some_and(|rel| {
-            rel.supports_nested_relationships
-                .as_ref()
-                .is_some_and(|n| n.supports_nested_in_ordering)
-        });
+    let orderable_object_type = object_types
+        .get(&order_by_expression.ordered_type)
+        .ok_or_else(|| models::ModelsError::OrderByExpressionError {
+            order_by_expression_identifier: order_by_expression.identifier.clone(),
+            error: OrderByExpressionError::UnknownOrderableType {
+                data_type: order_by_expression.ordered_type.clone(),
+            },
+        })?;
 
-    // If we have any relationships, and we don't support ordering by nested relationships, log an issue
     for relationship_name in order_by_expression.orderable_relationships.keys() {
-        if !connector_supports_ordering_nested_relationships {
-            issues.push(
-                models::ModelsIssue::OrderByExpressionContainsUnsupportedNestedRelationship {
-                    data_connector_name: model_source.data_connector.name.clone(),
-                    model_name: model_name.clone(),
+        // Get the relationship details
+        let relationship = orderable_object_type
+            .relationship_fields
+            .get(relationship_name)
+            .ok_or_else(|| models::ModelsError::OrderByExpressionError {
+                order_by_expression_identifier: order_by_expression.identifier.clone(),
+                error: OrderByExpressionError::UnknownRelationship {
                     relationship_name: relationship_name.clone(),
+                    object_type_name: order_by_expression.ordered_type.clone(),
+                },
+            })?;
+
+        // Get the target data connector name of the relationship, either from the target model or the target command
+        let target_data_connector_name = match &relationship.target {
+            object_relationships::RelationshipTarget::Model(model_relationship_target) => {
+                // Get the target model
+                let source = models
+                    .get(&model_relationship_target.model_name)
+                    .ok_or_else(|| models::ModelsError::ModelNotFound {
+                        model_name: model_relationship_target.model_name.clone(),
+                    })?
+                    .source
+                    .as_deref();
+
+                // If the model doesn't have a source, we need to raise an issue
+                if let Some(source) = source {
+                    &source.data_connector.name
+                } else {
+                    issues.push(
+                        models::ModelsIssue::OrderableRelationshipTargetModelMustHaveASource {
+                            order_by_expression_identifier: order_by_expression.identifier.clone(),
+                            model_name: model_name.clone(),
+                            relationship_name: relationship_name.clone(),
+                            target_model_name: model_relationship_target.model_name.clone(),
+                        }
+                        .into(),
+                    );
+                    continue;
+                }
+            }
+            object_relationships::RelationshipTarget::Command(command_relationship_target) => {
+                // Get the target command
+                let source = commands
+                    .get(&command_relationship_target.command_name)
+                    .ok_or_else(|| models::ModelsError::CommandNotFound {
+                        command_name: command_relationship_target.command_name.clone(),
+                    })?
+                    .source
+                    .as_deref();
+
+                // If the command doesn't have a source, we need to raise an issue
+                if let Some(source) = source {
+                    &source.data_connector.name
+                } else {
+                    issues.push(
+                        models::ModelsIssue::OrderableRelationshipTargetCommandMustHaveASource {
+                            order_by_expression_identifier: order_by_expression.identifier.clone(),
+                            model_name: model_name.clone(),
+                            relationship_name: relationship_name.clone(),
+                            target_command_name: command_relationship_target.command_name.clone(),
+                        }
+                        .into(),
+                    );
+                    continue;
+                }
+            }
+        };
+
+        // Validate the orderable relationship and log an issue if it fails
+        if let Err(error) = order_by_expressions::validate_orderable_relationship(
+            &order_by_expression.ordered_type,
+            relationship_name,
+            orderable_field_nestedness,
+            &model_source.data_connector,
+            target_data_connector_name,
+        ) {
+            issues.push(
+                models::ModelsIssue::OrderableRelationshipError {
                     order_by_expression_identifier: order_by_expression.identifier.clone(),
+                    model_name: model_name.clone(),
+                    error,
                 }
                 .into(),
             );
-        }
-    }
-
-    // If we have any orderable nested object fields, recur into them to see whether they have
-    // orderable relationships
-    for orderable_field in order_by_expression.orderable_fields.values() {
-        match orderable_field {
-            OrderableField::Object(orderable_object_field) => {
-                let nested_order_by_expression = order_by_expressions
-                    .objects
-                    .get(&orderable_object_field.order_by_expression_identifier)
-                    .ok_or_else(|| models::ModelsError::UnknownOrderByExpressionIdentifier {
-                        model_name: model_name.clone(),
-                        order_by_expression_identifier: orderable_object_field
-                            .order_by_expression_identifier
-                            .clone(),
-                    })?;
-                validate_nested_relationship_compatibility(
-                    nested_order_by_expression,
-                    model_source,
-                    model_name,
-                    order_by_expressions,
-                    issues,
-                )?;
-            }
-            OrderableField::Scalar(_) => {}
         }
     }