+++ v3/crates/metadata-resolve/src/stages/order_by_expressions/types.rs
@@ -131,3 +131,9 @@ impl ShouldBeAnError for OrderByExpressionIssue {
         }
     }
 }
+
+#[derive(Debug, Copy, Clone, Eq, PartialEq)]
+pub enum OrderableFieldNestedness {
+    NotNested,
+    ObjectNested,
+}